Implemented with database team
Moved database team changes over to this branch, implemented those functions for the socket connection
