+++ Backend/routes/messagesRoutes.js
@@ -1,7 +1,9 @@
 const express = require('express');
 const router = express.Router();
-const { markRead } = require('../controllers/messagesController')
+const { markRead, getAllMessages, addMessage } = require('../controllers/messagesController')
 
 router.post('/markRead', markRead);
+router.get('/getAllMessages', getAllMessages);
+router.post('/addMessage', addMessage);
 
 module.exports = router;
